move some headers to include/pluto so that they can be used in unit test seams
